pyproject.toml
[project]
name = "ai-travel-designer-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.4",
    "python-dotenv>=1.1.1",
]
